week_06/Day1/types_start_point/src/main/java/Bear.java
public class Bear {
    private String name;
    private int age;
    private String address;
    private double weight;

//    private int age;

//    if we need to add more properties then add to the top after private ie name and age

// this is to initialise the class
    public Bear(String name, int age, double weight){
        this.name = name;
        this.age = age;
        this.weight = weight;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public double getWeight() {
        return weight;
    }

    public void setWeight(double weight) {
        this.weight = weight;
    }

    public boolean readyToHibernate(){
        return this.weight > 80;
        }
    }
}
